Correction

+++ python_async_function/0-basic_async_syntax.py
@@ -3,14 +3,14 @@
                  (max_delay, with a default value of 10) named wait_random
                  that waits for a random delay between 0 and max_delay
                  (included and float value) seconds and eventually returns it
-    Arguments: max_delay: int = 10
+    Arguments: max_delay: float = 10.0
 '''
 
 import asyncio
 import random
 
 
-async def wait_random(max_delay: int = 10) -> float:
+async def wait_random(max_delay: float = 10.0) -> float:
     ''' Wait up to max_delay seconds and then return length of delay. '''
     delay = max_delay * random.random()
     await asyncio.sleep(delay)
